requirements.txt
requests
pytz
discord.py
rapidfuzz
pillow
word2number
websocket-client
python-dotenv
aiosqlite
Unidecode
aiohttp
aiofiles
lavalink
flashtext
dotenv